backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==4.0.0",
    "fastapi>=0.136.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.13.4",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.46.0",
]
